+++ CHANGELOG.md
@@ -11,6 +11,8 @@ All notable changes to this project will be documented in this file[^1].
 ## [2.41.0] - 2022-01-14
 ### Added
 - filter by gender to authorities
+- authority matching of swapped names
+- GMUHK harvest authority matching
 
 ### Changed
 - inscription length in item_translations table